Add an `upload` command
This change adds the ability to upload some Sightglass measurement data
to an ElasticSearch server. To do so, the parts of the data that are
environment-specific (machine, engine, benchmark) are "fingerprinted,"
or given a unique ID to differentiate them from other data points. The
"fingerprinting" also serves to unify data points more safely. E.g., if
the same benchmark is run on two different machines with the same
engine, the database will contain two machine entries and a single entry
each for the engine and benchmark; the measurement entries reference
these.

The mechanism for uploading the various data points to an HTTP endpoint
is specific to ElasticSearch in expectation of a future commit defining
how to run such a database. With some work, this `upload` functionality
could be made more generic or adapted to other database types.