ui: add a new Elasticsearch-based UI
This change replaces the now-removed custom UI with a vanilla
Elasticsearch + Kibana UI. The new `upload-elastic` command will push
data to the Elasticsearch document store and Kibana will display it with
various charts, visualizations, etc. The `ui/config` directory contains
some scripts to setup scripts to bring up the visualizations in a new
instance of the UI, deployed using `docker-compose`.